refactor(contract): move TEE verifier votes into api/tee_verifier.rs
